Correct bootstrap classes and alerts
To get us started the flash keys for alerts needed to be cleaned up so
that the account update and log in alerts would appear correctly in
bootstrap.  This problem still exists for the alerts that happen on the
Devise pages themselves in devise_error_messages!.

Cleaning up boostrap classes for the overall site will make it easier to
make unified changes to the site style.