Add :confirmable to Devise user model
Confirmable allows users to be confirmed via their e-mail and a
confirmation link.  Revised how the email looks and added the necessary
code to the migration file so that the existing users can be migrated
properly and are confirmed immediately.  Discovered the devise.en.yml
file which contains all of the localizations for Devise confirmation and
error messages, which will need to be revised to make sense for the
site.