Configure email and strings, test confirmation
Added a small test for confirmation.  Tests for the actual Grocerist
features are the only tests that should be added from here.  Devise
tests all their features heavily and testing their features is merely an
exercise in minitest rather than a useful test for the purposes of
actual application testing.  My response to the Devise controler
standard actions is what should be tested.

Also added email via SendGrid which needs to be tested in production.

Closes #5