Feature confirmable (#15)
* Add :confirmable to Devise user model

Confirmable allows users to be confirmed via their e-mail and a
confirmation link.  Revised how the email looks and added the necessary
code to the migration file so that the existing users can be migrated
properly and are confirmed immediately.  Discovered the devise.en.yml
file which contains all of the localizations for Devise confirmation and
error messages, which will need to be revised to make sense for the
site.

* Configure email and strings, test confirmation

Added a small test for confirmation.  Tests for the actual Grocerist
features are the only tests that should be added from here.  Devise
tests all their features heavily and testing their features is merely an
exercise in minitest rather than a useful test for the purposes of
actual application testing.  My response to the Devise controler
standard actions is what should be tested.

Also added email via SendGrid which needs to be tested in production.

Closes #5