Add delete mode button
To have a delete mode, we need a button to enter it.  The button
received some minor styling but for the most part inherits the
bootstrap button for now.  The button runs javascript to show all of the
delete buttons and change its' text.

Next steps:
- Allow the delete buttons to send AJAX requests with remote: :true
- Bind the successful AJAX response to the delete buttons to remove the
  item row from the page
- Modify the server code to support the new delete style

In the next feature we will have to preserve recently removed items so
that we can undo.