Finalize quickdelete feature
Added the ajax request to remove the need for a full redirect to see the
effect on the page elements.  Not sure if this removes compatibility
with non-javascript browsers but most likely it does.  Did some
tinkering with flash messages and managed to get them to render even
with the Ajax response, but that creates a new problem where leaving the
page preserves the flash past the first redirect since the ajax call
isn't a request/response that would empty the flash.  So this might need
to be logged as an issue.  Closes #17