Fix delete button appearing in an empty list (#22)
Fixed by not assigning @list_items if the items collection on the list
is empty.  Closes #19

What will happen when we inevitably want to add items to the list with
Ajax is uncertain.  Since adding the first item to the list will need to
render the delete button.  I think a decent way to do that will be to
make the delete button its' own partial and add it to the DOM using
jQuery, and render it from the server normally if there are already
items.